fix(ci): maintainability — split oversized verdict retry regressions
